Add 'optionalvalue' to flags tag.
Fields with 'flag' tag containing the optionalvalue will
attempt to parse any following argument into its type.
If successful, value is assigned to field, otherwise assumed to have no value.  With no value, the 'zero' type is assigned to the Field.

With maps or pointers, the pointer /map will be nil is the flag is not present at all.
When flag present, but no value, pointer map set to empty/zero value.
When flag value present, and parsable as the pointer type, value is assigned to the field